Fixes
- Fixed Responsive problem
- Fixed folder structure
- Fixed Bubble vectors behind animation on both login and signup pages
- Fixed width and some spacing
