+++ tailwind.config.js
@@ -9,7 +9,21 @@ module.exports = {
       tablets:{'max':'768px'},
       mobiles:{'max':'425px'},
       sScreens:{'max':'660px'},
-      sMobiles:{'max':'376px'}
+      sMobiles:{'max':'376px'},
+      'sm': '640px',
+      // => @media (min-width: 640px) { ... }
+
+      'md': '768px',
+      // => @media (min-width: 768px) { ... }
+
+      'lg': '1024px',
+      // => @media (min-width: 1024px) { ... }
+
+      'xl': '1280px',
+      // => @media (min-width: 1280px) { ... }
+
+      '2xl': '1536px',
+      // => @media (min-width: 1536px) { ... }
     },
     extend: {
       colors: {
